Making GTAP parsing faster
1)Missing function is improved to make it faster
2)Some functions removed from parser_gtap_mario_gdx into GTAP_utils.py
3)Added logs
4)Added exceptions
