+++ mario/tools/tableparser.py
@@ -47,6 +47,13 @@ from mario.tools.iomath import (
     calc_X_from_z,
 )
 
+from mario.tools.GTAP_utils import (
+    missing,
+    gdx_to_matrix,
+    gdx_to_matrix_rowname,
+    gdx_to_matrix_satellite,
+)
+
 import pandas as pd
 import logging
 import copy
@@ -1679,7 +1686,6 @@ def parser_figaro_e3(path,doping_value):
     EY = pd.DataFrame(0,index=E.index,columns=Y.columns)
     X = calc_X(Z, Y)
 
-
     #Doping Z because it is singular
     print(f"Replacing null values on Z diagional with {doping_value} to avoid singularity issues")
     for r in sets[_MASTER_INDEX['r']]:
@@ -1714,6 +1720,7 @@ def parser_figaro_e3(path,doping_value):
     return matrices, indeces, units
 
 def parser_gtap_mrio_csv(path):
+    #Not polished
     """
     Parse GTAP-MRIO tables in csv into MARIO format
     
@@ -2204,9 +2211,10 @@ def parser_gtap_mrio_gdx(path):
     tuple
         (matrices, indeces, units)
     """
-    from gams import transfer as gt
+    from gams import transfer as gt #to move outside
 
-    #Importing from the 5 GTAP csv data as dictionary
+    #Importing from the 5 GTAP gdx data tables as dictionary
+    log_time(logger, f"Parser: Reading tables from GDX files.")
     mrio_data = {}
     for key in _gtap_mrio_gdx_id.keys():
         file = _gtap_mrio_gdx_id[key]["file"]
@@ -2222,408 +2230,9 @@ def parser_gtap_mrio_gdx(path):
         "s": {"main": s}, #76 sectors
         #"k","f" later on
     }
-    
-    # Functions definition
-    def missing(df: pd.DataFrame,
-                variant: str,
-                indeces: dict,
-                ) -> pd.DataFrame:
-        #to optimize, takes too long
-        """
-        :param df: original DataFrame
-        :param variant: string that specifies the possible treatment:
-            - "general"
-            - "dom"
-            - "single_region"
-            - "single_region_va"
-            - "tax"
-            - "emi_dom"
-            - "emi_imp"
-            - "ene_dom"
-            - "ene_imp"
-        :param indeces: Dictionary of indices
 
-
-        Returns a DataFrame with missing values handled according to the selected mode.
-        """
-        # Extract useful lists
-        r = indeces['r']['main']
-        s = indeces['s']['main']
-        n = indeces['n']['main']
-
-        if variant == 'dom':
-            all_combinations = pd.MultiIndex.from_product([s, s+n, r], names=['COMM', 'agt', 'REG']).to_frame(index=False)
-
-            df_full = all_combinations.merge(df,on=['COMM','agt','REG'], how='left').fillna(0)
-            df_full['DST'] = df_full['REG'] #Filled only for same countryxcountry, block diagonal
-            df_full.rename(columns={'REG': 'SRC'}, inplace=True)
-            # Controllo dimensioni
-            expected_rows = len(s)*len(s+n)*len(r)
-            assert len(df_full) == expected_rows, f"[Variant Z] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-
-        elif variant == 'general': #slow
-            c = s 
-            a = s + n
-            all_combinations = pd.MultiIndex.from_product([c, a, r, r],names=['COMM', 'agt', 'SRC', 'DST']).to_frame(index=False)
-
-            df_full = all_combinations.merge(df,on=['COMM','agt','SRC','DST'],how='left').fillna(0)
-
-            expected_rows = len(c)*len(a)*(len(r)**2)
-            assert len(df_full) == expected_rows, f"[Variant items] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        elif variant == 'tax':
-            c = s
-            all_combinations = pd.MultiIndex.from_product([c, r, r],names=['COMM','SRC','DST']).to_frame(index=False)
-            df_full = all_combinations.merge(df,on=['COMM','SRC','DST'],how='left').fillna(0)
-
-            expected_rows = len(c)*len(r)*len(r)
-            assert len(df_full) == expected_rows, \
-                f"[Variant tax] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        elif variant == 'ptax':
-            c = s  
-            all_combinations = pd.MultiIndex.from_product([c, r],names=['COMM','REG']).to_frame(index=False)
-            df.drop(columns=['acts'],inplace=True)
-            df_full = all_combinations.merge(df,on=['COMM','REG'],how='left').fillna(0)
-
-            expected_rows = len(c)*len(r)
-            assert len(df_full) == expected_rows, \
-                f"[Variant tax] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        elif variant == 'single_region':
-            c = s  # oppure df['COMM'].unique()
-            a = s + n
-            all_combinations = pd.MultiIndex.from_product([c, a, r],names=['COMM','agt','DST']).to_frame(index=False)
-
-            df_full = all_combinations.merge(df,on=['COMM','agt','DST'],how='left').fillna(0)
-
-            expected_rows = len(c)*len(a)*len(r)
-
-            assert len(df_full) == expected_rows,f"[items_1reg] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        elif variant == 'single_region_va':
-            #when the commodities are the categories of value added
-            endw = df['ENDW'].unique().tolist()
-            all_combinations = pd.MultiIndex.from_product([endw, s, r],names=['ENDW','acts','DST']).to_frame(index=False)
-
-            df_full = all_combinations.merge(df,on=['ENDW','acts','DST'],how='left').fillna(0)
-
-            expected_rows = len(endw)*len(s)*len(r)
-            
-            assert len(df_full) == expected_rows,f"[items_1reg] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        elif variant == 'emi_dom':
-            #when the commodities are the categories of value added
-            df_full=pd.DataFrame(columns=['em','inputs','agt','SRC','DST','value'])
-            for e in df['em'].unique().tolist():
-                df_e=df[df['em']==e]
-                c = df_e['inputs'].unique().tolist()
-                a = s+n
-                all_combinations = pd.MultiIndex.from_product([c, a, r],names=['inputs','agt','SRC']).to_frame(index=False)
-                all_combinations['DST'] = all_combinations['SRC'] #Filled only for same countryxcountry, block diagonal
-                all_combinations['em'] = e
-                df_full_e = all_combinations.merge(df_e,on=['em','inputs','agt','SRC','DST'],how='left').fillna(0)
-                df_full=pd.concat([df_full,df_full_e],ignore_index=True)
-            return df_full
-        elif variant == 'emi_imp':
-            #when the commodities are the categories of value added
-            df_full=pd.DataFrame(columns=['em','inputs','agt','SRC','DST','value'])
-            for e in df['em'].unique().tolist():
-                df_e=df[df['em']==e]
-                c = df_e['inputs'].unique().tolist()
-                a = s+n
-                all_combinations = pd.MultiIndex.from_product([c,a,r,r],names=['inputs','agt','SRC','DST']).to_frame(index=False)
-                all_combinations['em'] = e
-                df_full_e = all_combinations.merge(df_e,on=['em','inputs','agt','SRC','DST'],how='left').fillna(0)
-                df_full=pd.concat([df_full,df_full_e],ignore_index=True)  
-            return df_full
-        elif variant == 'emi_proc':
-            #when the commodities are the categories of value added
-            df_full=pd.DataFrame(columns=['em','comm','acts','REG','value'])
-            for e in df['em'].unique().tolist():
-                df_e=df[df['em']==e]
-                c = df_e['comm'].unique().tolist()
-                a = s+n
-                all_combinations = pd.MultiIndex.from_product([c,a,r],names=['comm','acts','REG']).to_frame(index=False)
-                all_combinations['em'] = e
-                df_full_e = all_combinations.merge(df_e,on=['em','comm','acts','REG'],how='left').fillna(0)
-                df_full=pd.concat([df_full,df_full_e],ignore_index=True)  
-            return df_full
-
-        if variant == 'ene_dom':
-            c=df['ERG'].unique().tolist()
-            all_combinations = pd.MultiIndex.from_product([c, s+n, r], names=['ERG', 'agt', 'SRC']).to_frame(index=False)
-            all_combinations['DST'] = all_combinations['SRC'] #Filled only for same countryxcountry, block diagonal
-            
-            df_full = all_combinations.merge(df,on=['ERG','agt','SRC','DST'], how='left').fillna(0)
-            # Controllo dimensioni
-            expected_rows = len(c)*len(s+n)*len(r)
-            assert len(df_full) == expected_rows, f"[Variant energy dom] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        if variant == 'ene_imp':
-            c=df['ERG'].unique().tolist()
-            all_combinations = pd.MultiIndex.from_product([c, s+n, r,r], names=['ERG', 'agt', 'SRC','DST']).to_frame(index=False)
-
-            df_full = all_combinations.merge(df,on=['ERG','agt','SRC','DST'], how='left').fillna(0)
-            # Controllo dimensioni
-            expected_rows = len(c)*len(s+n)*(len(r)**2)
-            assert len(df_full) == expected_rows, f"[Variant energy imp] Expected {expected_rows} rows, got {len(df_full)}."
-            return df_full
-        
-        else:
-            raise ValueError(f"Unrecognized variant '{variant}'.")
-  
-        
-    def gdx_to_matrix(df: pd.DataFrame,
-                    var: str,
-                    variant_missing: str,
-                    indeces: dict,
-                    pivot_index: list = None,
-                    pivot_columns: list = None,
-                    ) -> pd.DataFrame or tuple:
-        """
-
-        1) Filters DataFrame through column 'VAR' == var to obtain one specific matrix.
-        2) Drop the 'VAR' column
-        3) Executes function missing() with appropriate variant.
-        4) If 'split_agt' is True, separates in Z and Y and has 2 pivot as output.
-        5) If 'split_agt' is False, unique pivot on pivot_index e pivot_columns.
-        
-        Returns:
-        - A DataFrame if split_agt = False
-        - A tuple (pivot_1, pivot_2) if split_agt = False
-        """
-
-        # 1) Filtrs according to var
-        if var is not None and (var in df.data.keys()):
-            df_filtered = df.data[var].records
-        else:
-            print("Error in filtering with 'VAR'.")
-        # 
-        # 2) Apply missing with desired variant
-        df_filled = missing(df_filtered, variant_missing, indeces)
-        
-        df_Z = df_filled[df_filled['agt'].isin(indeces['s']['main'])]
-        df_Y = df_filled[df_filled['agt'].isin(indeces['n']['main'])]
-
-        # 3) Pivoting
-        # Pivot sectorxsector section
-        pivot_Z = df_Z.pivot_table(
-            index=pivot_index, 
-            columns=pivot_columns, 
-            values='value', 
-            aggfunc='sum'
-        ).fillna(0)
-
-        # Pivot sectorxfinal demand section
-        pivot_Y = df_Y.pivot_table(
-            index=pivot_index, 
-            columns=pivot_columns, 
-            values='value', 
-            aggfunc='sum'
-        ).fillna(0)
-
-        return pivot_Z, pivot_Y
-
-    def gdx_to_matrix_rowname(df: pd.DataFrame,
-                    var: str,
-                    variant_missing: str,
-                    indeces: dict,
-                    row_name_setting: str,
-                    row_name_categ: str,
-                    row_name_reg: str = "",
-                    pivot_index: list = None,
-                    pivot_columns: list = None,
-                    split_agt: bool = False,
-                    ) -> pd.DataFrame or tuple:
-        """
-
-        1) Filters DataFrame through column 'VAR' == var to obtain one specific matrix.
-        2) Drop the 'VAR' column
-        3) Executes function missing() with appropriate variant.
-        4) If 'split_agt' is True, separates in Z and Y and has 2 pivot as output.
-        5) If 'split_agt' is False, unique pivot on pivot_index e pivot_columns.
-        
-        Returns:
-        - A DataFrame if split_agt = False
-        - A tuple (pivot_1, pivot_2) if split_agt = False
-        """
-
-        # 1) Filtrs according to var
-        if var is not None and (var in df.data.keys()):
-            df_filtered = df.data[var].records
-        else:
-            print("Error in filtering with 'VAR'.")
-
-        # 2) Apply missing with desired variant
-        df_filled = missing(df_filtered, variant_missing, indeces)
-
-        # 3) Change row name
-        if row_name_setting=='only_region':
-            df_filled['row_name']=row_name_categ+'_'+df_filled[row_name_reg]
-        elif row_name_setting=='only_categ':
-            df_filled['row_name']=row_name_categ+'_REG'
-        elif row_name_setting=='reg_comm':
-            df_filled['row_name']=row_name_categ+'_'+df_filled[row_name_reg]+'_'+df_filled['COMM']
-            df_filled.drop(columns='SRC',inplace=True)
-            df_filled.drop(columns='COMM',inplace=True)
-        elif row_name_setting=='only_endw':
-            df_filled['row_name']=row_name_categ+'_REG_'+df_filled['ENDW']
-        elif row_name_setting=='only_comm':
-            df_filled['row_name']=row_name_categ+'_REG_'+df_filled['COMM']
-
-
-        # 4) If need to split_agt
-        if split_agt:
-            df_Z = df_filled[df_filled['agt'].isin(indeces['s']['main'])]
-            df_Y = df_filled[df_filled['agt'].isin(indeces['n']['main'])]
-
-            # Pivot di part1
-            pivot_Z = df_Z.pivot_table(
-                index=pivot_index, 
-                columns=pivot_columns, 
-                values='value', 
-                aggfunc='sum'
-            ).fillna(0)
-
-            # Pivot di part2
-            pivot_Y = df_Y.pivot_table(
-                index=pivot_index, 
-                columns=pivot_columns, 
-                values='value', 
-                aggfunc='sum'
-            ).fillna(0)
-
-            return pivot_Z, pivot_Y
-
-        else:
-            # Pivot unico
-            matrix = df_filled.pivot_table(
-                index=pivot_index, 
-                columns=pivot_columns, 
-                values='value', 
-                aggfunc='sum'
-            ).fillna(0)
-
-            return matrix
-        
-    def gdx_to_matrix_satellite(df: pd.DataFrame,
-                    var: str,
-                    variant_missing: str,
-                    indeces: dict,
-                    row_name_setting: str,
-                    row_name_categ: str,
-                    pivot_index: list = None,
-                    pivot_columns: list = None,
-                    split_agt: bool = False,
-                    ) -> pd.DataFrame or tuple:
-        """
-
-        1) Filters DataFrame through column 'VAR' == var to obtain one specific matrix.
-        2) Drop the 'VAR' column
-        3) Executes function missing() with appropriate variant.
-        4) If 'split_agt' is True, separates in Z and Y and has 2 pivot as output.
-        5) If 'split_agt' is False, unique pivot on pivot_index e pivot_columns.
-        
-        Returns:
-        - A DataFrame if split_agt = False
-        - A tuple (pivot_1, pivot_2) if split_agt = False
-        """
-
-        # 1) Filtrs according to var
-        if var is not None and (var in df.data.keys()):
-            df_filtered = df.data[var].records
-        else:
-            print("Error in filtering with 'VAR'.")
-        
-        # 1b) If needed, select DOM/IMP
-        if row_name_setting=='emi_dom':
-            df_filtered=df_filtered[df_filtered['source']=='DOM'] 
-            df_filtered.drop(columns='source',inplace=True)
-        elif row_name_setting=='emi_imp':
-            df_filtered=df_filtered[df_filtered['source']=='IMP']
-            df_filtered.drop(columns='source',inplace=True)
-        if row_name_setting=='ene_dom':
-            df_filtered=df_filtered[df_filtered['SOURCE']=='DOM'] 
-            df_filtered.drop(columns='SOURCE',inplace=True)
-        elif row_name_setting=='ene_imp':
-            df_filtered=df_filtered[df_filtered['SOURCE']=='IMP']
-            df_filtered.drop(columns='SOURCE',inplace=True)
-        
-        # 2) Apply missing with desired variant
-        df_filled = missing(df_filtered, variant_missing, indeces)
-
-        # 3) Change row name
-        if row_name_setting=='emi_dom':
-            df_filled['row_name']=row_name_categ+'_'+df_filled['em']+'_dms_'+df_filled['inputs'] #careful: 'EM' element has different lengths
-            df_filled.drop(columns='em',inplace=True)
-            df_filled.drop(columns='SRC',inplace=True)
-            df_filled.drop(columns='inputs',inplace=True)
-        elif row_name_setting=='emi_imp':
-            df_filled['row_name']=row_name_categ+'_'+df_filled['em']+'_'+df_filled['SRC']+'_'+df_filled['inputs']
-            df_filled.drop(columns='em',inplace=True)
-            df_filled.drop(columns='SRC',inplace=True)
-            df_filled.drop(columns='inputs',inplace=True)
-        elif row_name_setting=='emi_proc':
-            df_filled['row_name']=row_name_categ+'_'+df_filled['em']+'_REG_'+df_filled['comm']
-            df_filled.drop(columns='em',inplace=True)
-            df_filled.drop(columns='comm',inplace=True)
-        elif row_name_setting=='ene_dom':
-            df_filled['row_name']=row_name_categ+'_dms_'+df_filled['ERG'] #domestic
-            df_filled.drop(columns='SRC',inplace=True)
-            df_filled.drop(columns='ERG',inplace=True)
-        elif row_name_setting=='ene_imp':
-            df_filled['row_name']=row_name_categ+'_'+df_filled['SRC']+'_'+df_filled['ERG']
-            df_filled.drop(columns='SRC',inplace=True)
-            df_filled.drop(columns='ERG',inplace=True)
-
-        # 4) If need to split_agt
-        if split_agt:
-            if row_name_setting=='emi_proc':
-                df_Z = df_filled[df_filled['acts'].isin(indeces['s']['main'])]
-                df_Y = df_filled[df_filled['acts'].isin(indeces['n']['main'])]
-            else:
-                df_Z = df_filled[df_filled['agt'].isin(indeces['s']['main'])]
-                df_Y = df_filled[df_filled['agt'].isin(indeces['n']['main'])]
-
-            # Pivot di part1
-            pivot_Z = df_Z.pivot_table(
-                index=pivot_index, 
-                columns=pivot_columns, 
-                values='value', 
-                aggfunc='sum'
-            ).fillna(0)
-
-            # Pivot di part2
-            pivot_Y = df_Y.pivot_table(
-                index=pivot_index, 
-                columns=pivot_columns, 
-                values='value', 
-                aggfunc='sum'
-            ).fillna(0)
-
-            return pivot_Z, pivot_Y
-
-        else:
-            # Pivot unico
-            matrix = df_filled.pivot_table(
-                index=pivot_index, 
-                columns=pivot_columns, 
-                values='value', 
-                aggfunc='sum'
-            ).fillna(0)
-
-            return matrix
-
-    # Matrix Z and Y
-    print('Starting: Z and Y')
+    #>>> Matrices Z and Y<<<
+    log_time(logger, f"Parsing Z and Y")
     Z_matrix_dom,Y_matrix_dom=gdx_to_matrix(mrio_data['SRCxDST'],'VDBA','dom',indeces,
                            pivot_index=['SRC','COMM'],pivot_columns=['DST','agt'])
     Z_matrix_imp,Y_matrix_imp=gdx_to_matrix(mrio_data['SRCxDST'],'VFOB','general',indeces,
@@ -2631,19 +2240,19 @@ def parser_gtap_mrio_gdx(path):
     Z=Z_matrix_dom+Z_matrix_imp
     Y=Y_matrix_dom+Y_matrix_imp
 
-    #Data from csv SRCxDST but to be included in V (long because of "general")
-    print('Starting: V')
+    #>>>Value added and taxes<<<
+    #Data from SRCxDST table but to be included in V and VY
+    log_time(logger, f"Parsing V and VY")
     V_mtax,VY_mtax=gdx_to_matrix_rowname(mrio_data['SRCxDST'],'MTAX','general',indeces,
                             split_agt=True,row_name_setting='reg_comm',row_name_categ='MTAX',row_name_reg='SRC',
                             pivot_index=['row_name'],pivot_columns=['DST','agt'])
     V_ittm,VY_ittm=gdx_to_matrix_rowname(mrio_data['SRCxDST'],'ITTM','general',indeces,
                                 split_agt=True,row_name_setting='reg_comm',row_name_categ='ITTM',row_name_reg='SRC',
                                 pivot_index=['row_name'],pivot_columns=['DST','agt'])
-    
     V=pd.concat([V_mtax,V_ittm],axis=0)
     VY=pd.concat([VY_mtax,VY_ittm],axis=0)
 
-    #Export and import taxes
+    # Export and import taxes from V-Tax table
     V_etax=gdx_to_matrix_rowname(mrio_data['V-Tax'],'ETAX','tax',indeces,
                            row_name_setting='only_region',row_name_categ='ETAX',row_name_reg='DST',
                            pivot_index=['row_name'],pivot_columns=['SRC','COMM'])
@@ -2653,7 +2262,7 @@ def parser_gtap_mrio_gdx(path):
 
     V=pd.concat([V,V_etax,V_ptax],axis=0)
 
-    #Value added and taxes from VA 
+    #Value added and taxes from VA table
     V_va=gdx_to_matrix_rowname(mrio_data['V'],'VA','single_region_va',indeces,row_name_setting='only_endw',
                            row_name_categ='VAAD',pivot_index=['row_name'],pivot_columns=['DST','acts'])
     V_vtax=gdx_to_matrix_rowname(mrio_data['V'],'VTAX','single_region_va',indeces,row_name_setting='only_endw',
@@ -2666,24 +2275,26 @@ def parser_gtap_mrio_gdx(path):
     V=pd.concat([V,V_va,V_vtax,V_idtax,V_imtax],axis=0)
     VY=pd.concat([VY,VY_idtax,VY_imtax],axis=0)
 
-    #Satellite account
-    # Filter the data for 'Emissions': Emi=from combustion Emi_proc=from processes
-    print("Stating: E and EY")
+    #>>>Environmental accounts E and EY<<<
+    #Filter the data for 'Emissions': Emi=from combustion Emi_proc=from processes
+    log_time(logger, f"Parsing E and EY")
     #Emissions from combustion
-    Emi_dom,Emi_dom_Y=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi','emi_dom',indeces,row_name_setting='emi_dom',
-                            row_name_categ='EMI',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
     Emi_dom_comb,Emi_dom_Y_comb=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi_COMB','emi_dom',indeces,row_name_setting='emi_dom',
                             row_name_categ='EMI',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
-    Emi_imp,Emi_imp_Y=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi','emi_imp',indeces,row_name_setting='emi_imp',
-                            row_name_categ='EMI',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
     Emi_imp_comb,Emi_imp_Y_comb=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi_COMB','emi_imp',indeces,row_name_setting='emi_imp',
+                            row_name_categ='EMI',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
+    
+    #Emissions general 
+    Emi_dom,Emi_dom_Y=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi','emi_dom',indeces,row_name_setting='emi_dom',
+                            row_name_categ='EMI',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
+    Emi_imp,Emi_imp_Y=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi','emi_imp',indeces,row_name_setting='emi_imp',
                             row_name_categ='EMI',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
     
     #Emissions from processes
     Emi_proc,Emi_proc_Y=gdx_to_matrix_satellite(mrio_data['Emissions'],'Emi_Proc','emi_proc',indeces,row_name_setting='emi_proc',
                             row_name_categ='E_P',split_agt=True,pivot_index=['row_name'],pivot_columns=['REG','acts'])
 
-    # Filter the data for 'E+EY - Energy'
+    # Filter the data for 'Energy'
     Ene_dom,Ene_dom_Y=gdx_to_matrix_satellite(mrio_data['Energy'],'NRG','ene_dom',indeces,row_name_setting='ene_dom',
                            row_name_categ='ENE',split_agt=True,pivot_index=['row_name'],pivot_columns=['DST','agt'])
     Ene_imp,Ene_imp_Y=gdx_to_matrix_satellite(mrio_data['Energy'],'NRG','ene_imp',indeces,row_name_setting='ene_imp',
@@ -2692,9 +2303,10 @@ def parser_gtap_mrio_gdx(path):
     E=pd.concat([Emi_dom,Emi_dom_comb,Emi_imp,Emi_imp_comb,Emi_proc,Ene_dom,Ene_imp],axis=0)
     EY=pd.concat([Emi_dom_Y,Emi_dom_Y_comb,Emi_imp_Y,Emi_imp_Y_comb,Emi_proc_Y,Ene_dom_Y,Ene_imp_Y],axis=0)
 
-    print("Finalizing")
+    #Finalizing
+    log_time(logger, f"Finalizing")
     indeces.update({
-        "k": {"main": E.index.tolist()}, #16743 = 37 emission categories, different commodities, domestic and 162x162 exchanges 'EMI_category_REG_COM' or 'ENE_REG_commodity' 
+        "k": {"main": E.index.tolist()}, #16843 = 37 emission categories, different commodities, domestic and 162x162 exchanges 'EMI_category_REG_COM' or 'ENE_REG_commodity' 
         "f": {"main": V.index.tolist()} #25116 Categories for Value added and Tax types 'CAT_REG_COM' characters: 3_3_variable
         })
 
@@ -2714,7 +2326,7 @@ def parser_gtap_mrio_gdx(path):
         index=indeces["f"]["main"],
         columns=["unit"],
     )
-
+    #Units for environmental extensions
     emi_series = ["M ton"] * sum(E.index.str[:3] == 'EMI')
     e_pi_series = ["M ton"] * sum(E.index.str[:3] == 'E_P')
     ene_series = ["M toe"] * sum(E.index.str[:3] == 'ENE')
@@ -2731,6 +2343,7 @@ def parser_gtap_mrio_gdx(path):
         _MASTER_INDEX["k"]: extensions_unit,
     }
 
+    #Preparing final dataframes
     Z.index = pd.MultiIndex.from_arrays([
         Z.index.get_level_values(0),
         [_MASTER_INDEX['s']]*Z.shape[0],
@@ -2747,8 +2360,8 @@ def parser_gtap_mrio_gdx(path):
     ])
     Y.index = Z.index
     EY.columns = Y.columns
+    VY.columns = Y.columns
 
-    # Calculate X matrix if not provided
     X = calc_X(Z, Y)
     X.index = Z.index
